Revert "[WIP] Reduce size of SerdNode header to 64 bits"
This reverts commit 857c8f3093b9c4877898ce074ec534f44f298a4f.

+++ src/node_internal.h
@@ -13,7 +13,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-static const size_t serd_node_align = 2 * sizeof(uint32_t);
+static const size_t serd_node_align = 2 * sizeof(uint64_t);
 
 /// Return the total size in bytes of a node with the given string length
 ZIX_CONST_FUNC size_t
